Add Create OneNote Link from Clipboard
Added macro to create a (source) style link in OneNote from the
Clipboard.

Created this macro in response to frequent need to cite YouTube videos
at specific points in the video in OneNote.  Looked at automating
Chrome with AutoHotkey - chrome.ahk gets recommended often - but this
will do the job for now.