Add Copy URL with Time to OneNote from YouTube
Add AutoHotkey macro to copy a URL with time from a YouTube video in
Chrome to OneNote.

This macro makes use of the YouTube feature "Copy video URL at current
Time" to create a link in OneNote that references a specific point in a
YouTube video.

By creating a link to a specific point in the video, the OneNote link
becomes much more powerfull - no more scanning/watching the video to
find the referenced material.