Automate App Switching
Refactored to automate the app switch from OneNote to Chrome. This means
this macro shortcut can now be activated from OneNote. Previously, user
was required to manually switch to Chrome from OneNote with alt+tab or
manual mouse movement and click or similar.

This manual step was intentional in that it was thought that it
would prevent a shortcut being copied from the wrong browser, if more
than one was open. There was however, a negative externality in that it
decreased the efficiency of this macro. With this refactor, the
efficiency has been restored. However, this refactor may have an
negative, unintended consequence in there is more than one instance of
Chrome open. In testing, it's the last used instance of Chrome that gets
 activated, which seems appropriate, but more live testing needed to
determine if currrent tests reflect actual usage.