Refactor Macro to Use New Functions
Refactored the
^!Numpad1::
^!F1::
AutoHotKey macro, which copys a URL from the Chrome Omnibox and creates
a (source) style link in OneNote, to use new functions.

The purpose of this change is to work towards the DRY principle,
by reducing code duplication.

Made small changes to fix typos, etc. in line with campsite principle.