Remove Macro for Creating (source) Link in OneNote
Removed the macro with HotKeys:
^!Numpad5::
^!F5::
which was used to created a (source) link in OneNote from a link
saved on the clipboard by the "Copy video URL at current time" YouTube
feature.

This macro was deprecated because another macro was created in which
the generation of the link on the YouTube-side was automated.