BugFix: Fix Send {RAW} Error
This change fixes a bug that was introduced when refactoring SendRaw to
Send {Raw}.

The bug was instroduced by the space between {Raw} and %URL% in this
command: Send {Raw} %URL%

The space between {Raw} and %URL% causes a space to be printed
immediately before the URL. This in turn prevents the creation of the
link in OneNote.

To function as expected the command needs to look like this:
    Send {Raw}%URL%
Note that there is no space between {Raw} and %URL%.

This commit also instroduces a HotString correcting SHA-1 proper case
issues.