BugFix: Fix HotString ahk
HotString for AutoHotKey (ahk) was not working properly when used in the
Chrome OmniBox.  Executing the ahk HotString in the Chrome OmniBox
frequently resulted in an additional 'a' being pre-pended to the
replaced text, i.e., aAutoHotKey resulted instead of AutoHotKey.

The behavior of the HotString was more predictable outside of Chrome
however, the switch from a basic HotString to Send necessitated
the inclusion of a {SPACE} to the Send command so double-tapping the
space bar wasn't required. Without the inclusion of the {SPACE} the
first time the space bar was hit would initiate the auto-replacement but
then another tap on the space bar was required to add the expected
space after the auto-replaced text.