Make Paste Text Only Macro Context-Sensitive
Refactored the paste text only macro (^+v::) to be context-sensitive so
that it will only fire if OneNote is active. This prevents this macro
from attempting to perform a text only paste in a app that doesn't
support it the way OneNote does.

Refactored macro to write Atlassian noformat block (#[::). Macro was
failing #Warn. Macro appeared to be using outdated AutoHotkey syntax.

Refactored macro to write Atlassian SQL block (#]::). Macro was
failing #Warn. Macro appeared to be using outdated AutoHotkey syntax.

Note: Refactored the Atlassian noformat and SQL block scripts because
they were causing #Warn errors. The two macros haven't been used in
quite some time and are ripe for deprecating or updating if Atlassian
use is resumed. For now, this is strike one on these two Atlassian
macros.