Restored ^!F5 Macro: Create Link from clipboard.
Restored the ^!F5 macro which creates a OneNote (source)-style link
from the clipboard. This macro was an incremental step in the
development of the ^!F5 macro and was deleted when the ^!F5 macro was
created. However, it turns out this macro is useful.

The ^!F5 macro has been restored as it was before the delete. However,
this macro (^!F5) needs to be refactored to make use of the functions
and needs to be made context-sensitive since this macro is only for
use with OneNote.