Make ^!F6 Context-Sensitive. Fix URL Regex
Refactored the ^!F6 OneNote macro for creating a (source)-style link
from a YouTube video with timestamp to be context-sensitive. This macro
is specific to OneNote so shouldn't fire unless very specific conditions
are met. See implmentation for details.

Fixed an edge-case bug with timestamped URLs coming from YouTube where
the URL contains an underscore (_). This case was not present in the
regex to validate the URLs, but is now handled with this fix.